backend/requirements.txt
# ============================================
# SatyaSetu.AI Backend - Complete Dependencies
# ============================================

# --- Core Framework ---
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
pydantic>=2.0.0

# --- OCR & Image Processing ---
pytesseract>=0.3.10
Pillow>=10.0.0
opencv-python>=4.8.0
numpy>=1.24.0
PyMuPDF>=1.23.0  # provides 'fitz' for PDF OCR

# --- NLP & Text Processing ---
spacy>=3.7.0
textblob>=0.17.1

# --- Machine Learning ---
scikit-learn>=1.3.0
sentence-transformers>=2.2.0
joblib>=1.3.0
xgboost>=2.0.0

# --- PDF Report Generation ---
reportlab>=4.0.0
fpdf2>=2.7.0
matplotlib>=3.7.0

# --- HTTP & OSINT ---
requests>=2.31.0
aiohttp>=3.8.0

# --- Database ---
psycopg2-binary>=2.9.0

# --- AI Copilot ---
groq>=1.0.0

# --- Utilities ---
tqdm>=4.65.0

pandas>=1.5

# OCR Dependencies
easyocr
opencv-python-headless

# NER Dependencies
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz

# Scam Classifier Dependencies
sentence-transformers
scikit-learn
textblob
joblib